pyproject.toml
[project]
name = "airgap-llm-edu"
version = "0.1.0"
description = "폐쇄망 LLM 실무 2일 교육 과정 — LangChain·RAG·GraphRAG·보안·OpenWebUI"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "airgap-llm-edu" }]

dependencies = [
    # Core LangChain
    "langchain>=0.3.19,<0.4",
    "langchain-core>=0.3.40,<0.4",
    "langchain-community>=0.3.18,<0.4",
    "langchain-openai>=0.3.7,<0.4",
    "langchain-text-splitters>=0.3.6,<0.4",
    "langchain-experimental>=0.3.4,<0.4",
    "langchain-huggingface>=0.1.2,<0.2",
    "langgraph>=0.2.76,<0.3",
    "langsmith>=0.3.11,<0.4",
    # OpenAI direct + tokenization
    "openai>=1.66.3,<2",
    "tiktoken>=0.9.0,<0.10",
    # Local / comparison embeddings
    "sentence-transformers>=3.4.1,<4",
    "FlagEmbedding>=1.3.4,<2",
    # Vector stores
    "chromadb>=0.6.3,<0.7",
    "faiss-cpu>=1.10.0,<2",
    "qdrant-client>=1.13.2,<2",
    "rank-bm25>=0.2.2,<0.3",
    # Document loaders / splitters
    "pypdf>=5.3.0",
    "pymupdf>=1.25.3",
    "pymupdf4llm>=0.0.17",
    "unstructured[pdf]>=0.16.16",
    "python-docx>=1.1.2",
    "docx2txt>=0.8",
    "python-hwpx>=0.1.3",
    "kss>=6.0.4",
    "kiwipiepy>=0.20.3",
    # Evaluation
    "ragas>=0.2.13,<0.3",
    "datasets>=3.3.2",
    # Graph RAG
    "neo4j>=5.28.1",
    "langchain-neo4j>=0.3.0",
    # Web service (S4-4)
    "fastapi>=0.115.11",
    "uvicorn[standard]>=0.34.0",
    "gradio>=5.16.1",
    "sse-starlette>=2.2.1",
    # Notebook UX
    "matplotlib>=3.10.1",
    "seaborn>=0.13.2",
    "jupyter>=1.1.1",
    "jupyterlab>=4.3.5",
    "ipywidgets>=8.1.5",
    "bertviz>=1.4.0",
    "nest-asyncio>=1.6.0",
    # Security / guardrails
    "presidio-analyzer>=2.2.357",
    "presidio-anonymizer>=2.2.357",
    "nemoguardrails>=0.12.0",
    # Utils
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.6",
    "httpx>=0.28.1",
    "tqdm>=4.67.1",
    "pyyaml>=6.0.2",
    "langchain-qdrant>=0.2.1",
    "streamlit>=1.56.0",
    "ddgs>=9.14.0",
    "numpy>=2.4.4",
    "pandas>=2.3.3",
    "torch>=2.11.0",
    "transformers>=4.57.6",
]

[tool.uv]
# 폐쇄망 반입 시 source distribution보다 wheel을 선호
prerelease = "if-necessary-or-explicit"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["common"]
